Unified common code in action-list.ts and equipment-list.ts in new item-list.ts file
